Add example sources for easier validation
This commit adds the complete example files from the spec, so that they
can be individually loaded and validated. This doesn’t help with
synchronising changes between the HTML document and the examples of
course, so it’s just a handy reference material. Arguably they should
have the .ttml extension since that’s the registered type extension,
but xml is not exactly wrong and is more easily processed by generic
XML processors.